refactor: reorganize src/ into DDD bounded contexts
Move flat module layout into three directories reflecting domain
architecture: engine/ (chess logic), audio/ (synthesis pipeline),
tui/ (display and REPL). Pure structural change — zero logic
modifications, 157/157 tests passing.
